backend/requirements.txt
fastapi
uvicorn
pydantic
pysentimiento
pytest
pytest-asyncio
httpx
PyJWT
passlib[bcrypt]